64-bit PatchInterface
We have an abstract 'PatchInterface' class that efficiently allows to have a
patched view on a file without having to load that whole file in memory (e.g.
this is useful for patched hardisk images and to a lesser degree also for
diskimages).

This patch extends PatchInterface to work with 64-bit underlying sizes.
Although the only concrete patch fileformat (IPS) that we currently support is
limited to 24-bit offsets.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13087 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/EmptyPatch.hh
@@ -11,16 +11,16 @@ namespace openmsx {
 class EmptyPatch : public PatchInterface, private noncopyable
 {
 public:
-	EmptyPatch(const byte* block, unsigned size);
+	EmptyPatch(const byte* block, size_t size);
 
-	virtual void copyBlock(unsigned src, byte* dst, unsigned num) const;
-	virtual unsigned getSize() const;
+	virtual void copyBlock(size_t src, byte* dst, size_t num) const;
+	virtual size_t getSize() const;
 	virtual std::vector<Filename> getFilenames() const;
 	virtual bool isEmptyPatch() const { return true; }
 
 private:
 	const byte* block;
-	const unsigned size;
+	const size_t size;
 };
 
 } // namespace openmsx
